// crm-sales-hub/src/generated/models/OpportunitiesModel.ts
export const Opportunitiesbudgetstatus = {
  0: 'NoCommittedBudget',
  1: 'MayBuy',
  2: 'CanBuy',
  3: 'WillBuy'
} as const;
export type Opportunitiesbudgetstatus = keyof typeof Opportunitiesbudgetstatus;
export const Opportunitiescaptureproposalfeedback = {
  0: 'Yes',
  1: 'No'
} as const;
export type Opportunitiescaptureproposalfeedback = keyof typeof Opportunitiescaptureproposalfeedback;
export const Opportunitiescompletefinalproposal = {
  0: 'Markcomplete',
  1: 'Completed'
} as const;
export type Opportunitiescompletefinalproposal = keyof typeof Opportunitiescompletefinalproposal;
export const Opportunitiescompleteinternalreview = {
  0: 'Markcomplete',
  1: 'Completed'
} as const;
export type Opportunitiescompleteinternalreview = keyof typeof Opportunitiescompleteinternalreview;
export const Opportunitiesconfirminterest = {
  0: 'Yes',
  1: 'No'
} as const;
export type Opportunitiesconfirminterest = keyof typeof Opportunitiesconfirminterest;
export const Opportunitiesdecisionmaker = {
  0: 'markcomplete',
  1: 'completed'
} as const;
export type Opportunitiesdecisionmaker = keyof typeof Opportunitiesdecisionmaker;
export const Opportunitiesdevelopproposal = {
  0: 'Markcomplete',
  1: 'Completed'
} as const;
export type Opportunitiesdevelopproposal = keyof typeof Opportunitiesdevelopproposal;
export const Opportunitiesevaluatefit = {
  0: 'Yes',
  1: 'No'
} as const;
export type Opportunitiesevaluatefit = keyof typeof Opportunitiesevaluatefit;
export const Opportunitiesfiledebrief = {
  0: 'Markcomplete',
  1: 'Completed'
} as const;
export type Opportunitiesfiledebrief = keyof typeof Opportunitiesfiledebrief;
export const Opportunitiesidentifycompetitors = {
  0: 'markcomplete',
  1: 'completed'
} as const;
export type Opportunitiesidentifycompetitors = keyof typeof Opportunitiesidentifycompetitors;
export const Opportunitiesidentifycustomercontacts = {
  0: 'markcomplete',
  1: 'completed'
} as const;
export type Opportunitiesidentifycustomercontacts = keyof typeof Opportunitiesidentifycustomercontacts;
export const Opportunitiesidentifypursuitteam = {
  0: 'markcomplete',
  1: 'completed'
} as const;
export type Opportunitiesidentifypursuitteam = keyof typeof Opportunitiesidentifypursuitteam;
export const Opportunitiesinitialcommunication = {
  0: 'Contacted',
  1: 'NotContacted'
} as const;
export type Opportunitiesinitialcommunication = keyof typeof Opportunitiesinitialcommunication;
export const Opportunitiesisprivate = {
  0: 'No',
  1: 'Yes'
} as const;
export type Opportunitiesisprivate = keyof typeof Opportunitiesisprivate;
export const Opportunitiesisrevenuesystemcalculated = {
  0: 'UserProvided',
  1: 'SystemCalculated'
} as const;
export type Opportunitiesisrevenuesystemcalculated = keyof typeof Opportunitiesisrevenuesystemcalculated;
export const Opportunitiesmsdyn_forecastcategory = {
  100000001: 'Pipeline',
  100000002: 'Bestcase',
  100000003: 'Committed',
  100000004: 'Omitted',
  100000005: 'Won',
  100000006: 'Lost'
} as const;
export type Opportunitiesmsdyn_forecastcategory = keyof typeof Opportunitiesmsdyn_forecastcategory;
export const Opportunitiesmsdyn_gdproptout = {
  0: 'No',
  1: 'Yes'
} as const;
export type Opportunitiesmsdyn_gdproptout = keyof typeof Opportunitiesmsdyn_gdproptout;
export const Opportunitiesmsdyn_opportunitygrade = {
  0: 'GradeA',
  1: 'GradeB',
  2: 'GradeC',
  3: 'GradeD'
} as const;
export type Opportunitiesmsdyn_opportunitygrade = keyof typeof Opportunitiesmsdyn_opportunitygrade;
export const Opportunitiesmsdyn_opportunityscoretrend = {
  0: 'Improving',
  1: 'Steady',
  2: 'Declining',
  3: 'Notenoughinfo'
} as const;
export type Opportunitiesmsdyn_opportunityscoretrend = keyof typeof Opportunitiesmsdyn_opportunityscoretrend;
export const Opportunitiesmsdyn_ordertype = {
  192350000: 'Itembased',
  690970002: 'Service_MaintenanceBased'
} as const;
export type Opportunitiesmsdyn_ordertype = keyof typeof Opportunitiesmsdyn_ordertype;
export const Opportunitiesneed = {
  0: 'Musthave',
  1: 'Shouldhave',
  2: 'Goodtohave',
  3: 'Noneed'
} as const;
export type Opportunitiesneed = keyof typeof Opportunitiesneed;
// ...
export interface OpportunitiesBase {
  actualclosedate?: string;
  actualvalue?: string;
  budgetamount?: string;
  budgetstatus?: Opportunitiesbudgetstatus;
  "CampaignId@odata.bind"?: string;
  captureproposalfeedback?: Opportunitiescaptureproposalfeedback;
  closeprobability?: string;
  completefinalproposal?: Opportunitiescompletefinalproposal;
  completeinternalreview?: Opportunitiescompleteinternalreview;
  confirminterest?: Opportunitiesconfirminterest;
  currentsituation?: string;
  customerid: string;
  customeridtype: string;
  customerneed?: string;
  customerpainpoints?: string;
  decisionmaker?: Opportunitiesdecisionmaker;
  description?: string;
  developproposal?: Opportunitiesdevelopproposal;
  discountamount?: string;
  discountpercentage?: string;
  emailaddress?: string;
  estimatedclosedate?: string;
  estimatedvalue?: string;
  evaluatefit?: Opportunitiesevaluatefit;
  filedebrief?: Opportunitiesfiledebrief;
  finaldecisiondate?: string;
  freightamount?: string;
  identifycompetitors?: Opportunitiesidentifycompetitors;
  identifycustomercontacts?: Opportunitiesidentifycustomercontacts;
  identifypursuitteam?: Opportunitiesidentifypursuitteam;
  importsequencenumber?: string;
  initialcommunication?: Opportunitiesinitialcommunication;
  isrevenuesystemcalculated?: Opportunitiesisrevenuesystemcalculated;
  lastonholdtime?: string;
  msdyn_forecastcategory?: Opportunitiesmsdyn_forecastcategory;
  msdyn_gdproptout?: Opportunitiesmsdyn_gdproptout;
  msdyn_opportunitygrade?: Opportunitiesmsdyn_opportunitygrade;
  "msdyn_opportunitykpiid@odata.bind"?: string;
  msdyn_opportunityscore?: string;
  msdyn_opportunityscoretrend?: Opportunitiesmsdyn_opportunityscoretrend;
  msdyn_ordertype?: Opportunitiesmsdyn_ordertype;
  "msdyn_PredictiveScoreId@odata.bind"?: string;
  msdyn_scorehistory?: string;
  msdyn_scorereasons?: string;
  "msdyn_segmentid@odata.bind"?: string;
  msdyn_similaropportunities?: string;
  "msdyn_WorkOrderType@odata.bind"?: string;
  name: string;
  need?: Opportunitiesneed;
  opportunityid: string;
  opportunityratingcode?: Opportunitiesopportunityratingcode;
  "OriginatingLeadId@odata.bind"?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  "ParentAccountId@odata.bind"?: string;
  "ParentContactId@odata.bind"?: string;
  participatesinworkflow?: Opportunitiesparticipatesinworkflow;
  presentfinalproposal?: Opportunitiespresentfinalproposal;
  presentproposal?: Opportunitiespresentproposal;
  "PriceLevelId@odata.bind"?: string;
  pricingerrorcode?: Opportunitiespricingerrorcode;
  prioritycode?: Opportunitiesprioritycode;
  processid?: string;
  proposedsolution?: string;
  purchaseprocess?: Opportunitiespurchaseprocess;
  purchasetimeframe?: Opportunitiespurchasetimeframe;
  pursuitdecision?: Opportunitiespursuitdecision;
  qualificationcomments?: string;
  quotecomments?: string;
  resolvefeedback?: Opportunitiesresolvefeedback;
  salesstage?: Opportunitiessalesstage;
  salesstagecode?: Opportunitiessalesstagecode;
  schedulefollowup_prospect?: string;
  schedulefollowup_qualify?: string;
  scheduleproposalmeeting?: string;
  sendthankyounote?: Opportunitiessendthankyounote;
  skippricecalculation?: Opportunitiesskippricecalculation;
  "SLAId@odata.bind"?: string;
  stageid?: string;
  statecode: Opportunitiesstatecode;
  statuscode?: string;
  stepid?: string;
  stepname?: string;
  teamsfollowed?: string;
  timeline?: Opportunitiestimeline;
  timezoneruleversionnumber?: string;
  totalamount?: string;
  totalamountlessfreight?: string;
  totaldiscountamount?: string;
  totallineitemamount?: string;
  totallineitemdiscountamount?: string;
  totaltax?: string;
  "TransactionCurrencyId@odata.bind": string;
  traversedpath?: string;
  utcconversiontimezonecode?: string;
}

export interface Opportunities extends OpportunitiesBase {
  accountidname?: string;
  accountidyominame?: string;
  actualvalue_base?: string;
  budgetamount_base?: string;
  budgettypename?: string;
  campaignidname?: string;
  captureproposalfeedbackname?: string;
  completefinalproposalname?: string;
  completeinternalreviewname?: string;
  confirminterestname?: string;
  contactidname?: string;
  contactidyominame?: string;
  createdbyname?: string;
  createdbyyominame?: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame?: string;
  customeridname: string;
  customeridyominame: string;
  decisionmakername?: string;
  developproposalname?: string;
  discountamount_base?: string;
  estimatedvalue_base?: string;
  evaluatefitname?: string;
  exchangerate?: string;
  filedebriefname?: string;
  freightamount_base?: string;
  identifycompetitorsname?: string;
  identifycustomercontactsname?: string;
  identifypursuitteamname?: string;
  initialcommunicationname?: string;
  isprivate?: Opportunitiesisprivate;
  isprivatename?: string;
  isrevenuesystemcalculatedname?: string;
  modifiedbyname?: string;
  modifiedbyyominame?: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame?: string;
  msdyn_forecastcategoryname?: string;
  msdyn_gdproptoutname?: string;
  msdyn_opportunitygradename?: string;
  msdyn_opportunitykpiidname?: string;
  msdyn_opportunityscoretrendname?: string;
  msdyn_ordertypename?: string;
  msdyn_predictivescoreidname?: string;
  msdyn_segmentidname?: string;
  msdyn_workordertypename?: string;
  needname?: string;
  onholdtime?: string;
  opportunityratingcodename?: string;
  originatingleadidname?: string;
  originatingleadidyominame?: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  parentaccountidname?: string;
  parentaccountidyominame?: string;
  parentcontactidname?: string;
  parentcontactidyominame?: string;
  participatesinworkflowname?: string;
  presentfinalproposalname?: string;
  presentproposalname?: string;
  pricelevelidname: string;
  pricingerrorcodename?: string;
  prioritycodename?: string;
  purchaseprocessname?: string;
  purchasetimeframename?: string;
  pursuitdecisionname?: string;
  resolvefeedbackname?: string;
  salesstagecodename?: string;
  salesstagename?: string;
  sendthankyounotename?: string;
  skippricecalculationname?: string;
  slainvokedidname?: string;
  slaname?: string;
  statecodename?: string;
  statuscodename?: string;
  timelinename?: string;
  timespentbymeonemailandmeetings?: string;
  totalamount_base?: string;
  totalamountlessfreight_base?: string;
  totaldiscountamount_base?: string;
  totallineitemamount_base?: string;
  totallineitemdiscountamount_base?: string;
  totaltax_base?: string;
  transactioncurrencyidname?: string;
  versionnumber?: string;
  accountid?: object;
  _accountid_value?: string;
  campaignid?: object;
  _campaignid_value?: string;
  contactid?: object;
  _contactid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  msdyn_opportunitykpiid?: object;
  _msdyn_opportunitykpiid_value?: string;
  msdyn_predictivescoreid?: object;
  _msdyn_predictivescoreid_value?: string;
  msdyn_segmentid?: object;
  _msdyn_segmentid_value?: string;
  msdyn_workordertype?: object;
  _msdyn_workordertype_value?: string;
  originatingleadid?: object;
  _originatingleadid_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  parentaccountid?: object;
  _parentaccountid_value?: string;
  parentcontactid?: object;
  _parentcontactid_value?: string;
  pricelevelid?: object;
  _pricelevelid_value?: string;
  slaid?: object;
  _slaid_value?: string;
  slainvokedid?: object;
  _slainvokedid_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}